Add files via upload
Input a number by a user, checks to see if the number you chose matches the number the computer chose at random. If it is the same it will ask you if you want to play again. If you guess more than 5 times, and you still have it wrong, it will end the game and ask you if you want to play again. If you guess it right, it will congratulate you. Also, this prints out the guesses you have already tried.